Update http.rb
The streaming proc method signature changed in Faraday 2.5.0, and any older version only provides the first two arguments to the proc ([see 6799f58](https://github.com/lostisland/faraday/commit/6799f5852d31dae32699949790633c68282ab71d)).

If using an older version of Faraday, we don't have a way to know if the streaming response is successful, so we just assume it is.