Update README.md
I noticed you wrote both the openai and anthropic clients, and realized that the open-ai client is very similar to the anthropic client but the anthropic readme doesn't document all the functionality. Figured I would add this to the readme to let others know this functionality exists on the anthropic client as well.